Replace activity on history stack on older Android versions
to prevent the need to press back multiple times to go back on activity stack.
This issue only appears on older Android versions. Somewhere before Android 9.
When an activity starts itself, it does not remove itself from the history stack. This however can be enforced by providing flags to the starting intent.