Pr 17 replay (#19)
* Fix/Refactor language support

The previous language support code has a long history in code and in bugs.
This refactor attempts to get rid of all remaining bugs and clean up the code to make it more readable.

Old issues:
- Crash when unsupported system language. (Already fixed on here)
- On orientation change, it switches to English.
- When language was fixed, it causes a loop on orientation change.
- One activity never bothered to restart (check for language changes).

The LanguageSetting class has been removed and the enum LanguageCode became its own file.
There is no need to store system and language code together in an activity. That only adds complexity and weird cases.
Now EACH activity stores its language code and checks whenever it gets opened/turned if the locale is still correct.

onConfigurationChange never seemed to be called. Its purpose was not obvious to me either. I removed all which had been added for language.

* fix build (my merge broke it)

* Replace activity on history stack on older Android versions

to prevent the need to press back multiple times to go back on activity stack.
This issue only appears on older Android versions. Somewhere before Android 9.
When an activity starts itself, it does not remove itself from the history stack. This however can be enforced by providing flags to the starting intent.

Co-authored-by: Ecconia <public-commit@ecconia.com>